requirements.txt
langchain>=0.1.0
langchain-core>=0.1.20
langchain-community>=0.0.12
langgraph>=0.0.20
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
moviepy>=1.0.3
jinja2>=3.1.2
replicate>=0.25.0
chromadb>=0.4.24
elevenlabs>=1.0.0
